Move legacy codecs to backward_codecs
