+++ CHANGELOG.md
@@ -22,4 +22,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Maintenance
 * Update `com.google.guava:failureaccess` from 1.0.1 to 1.0.2
 * Update `com.google.guava:guava` from 32.1.3-jre to 33.2.1-jre
+* Move legacy codecs to backward_codecs [389](https://github.com/opensearch-project/opensearch-jvector/pull/389)
 ### Refactoring
