+++ src/main/java/org/opensearch/knn/index/codec/backward_codecs/KNN910Codec/KNN910Codec.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.knn.index.codec.KNN910Codec;
+package org.opensearch.knn.index.codec.backward_codecs.KNN910Codec;
 
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.codecs.CompoundFormat;
