+++ src/main/java/org/opensearch/knn/index/codec/backward_codecs/KNN990Codec/KNN990Codec.java
@@ -2,10 +2,10 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.knn.index.codec.KNN920Codec;
+
+package org.opensearch.knn.index.codec.backward_codecs.KNN990Codec;
 
 import lombok.Builder;
-import lombok.extern.log4j.Log4j2;
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.codecs.CompoundFormat;
 import org.apache.lucene.codecs.DocValuesFormat;
@@ -16,29 +16,29 @@ import org.opensearch.knn.index.codec.KNNCodecVersion;
 import org.opensearch.knn.index.codec.KNNFormatFacade;
 
 /**
- * KNN codec that is based on Lucene92 codec
+ * KNN Codec that wraps the Lucene Codec which is part of Lucene 9.9
  */
-@Log4j2
-public final class KNN920Codec extends FilterCodec {
-    private static final KNNCodecVersion VERSION = KNNCodecVersion.V_9_2_0;
+public class KNN990Codec extends FilterCodec {
+    private static final KNNCodecVersion VERSION = KNNCodecVersion.V_9_9_0;
     private final KNNFormatFacade knnFormatFacade;
     private final PerFieldKnnVectorsFormat perFieldKnnVectorsFormat;
 
     /**
-     * No arg constructor that uses Lucene91 as the delegate
+     * No arg constructor that uses Lucene99 as the delegate
      */
-    public KNN920Codec() {
+    public KNN990Codec() {
         this(VERSION.getDefaultCodecDelegate(), VERSION.getPerFieldKnnVectorsFormat());
     }
 
     /**
-     * Constructor that takes a Codec delegate to delegate all methods this code does not implement to.
+     * Sole constructor. When subclassing this codec, create a no-arg ctor and pass the delegate codec
+     * and a unique name to this ctor.
      *
      * @param delegate codec that will perform all operations this codec does not override
      * @param knnVectorsFormat per field format for KnnVector
      */
     @Builder
-    public KNN920Codec(Codec delegate, PerFieldKnnVectorsFormat knnVectorsFormat) {
+    protected KNN990Codec(Codec delegate, PerFieldKnnVectorsFormat knnVectorsFormat) {
         super(VERSION.getCodecName(), delegate);
         knnFormatFacade = VERSION.getKnnFormatFacadeSupplier().apply(delegate);
         perFieldKnnVectorsFormat = knnVectorsFormat;
