+++ src/main/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -1,8 +1,8 @@
-org.opensearch.knn.index.codec.KNN910Codec.KNN910Codec
-org.opensearch.knn.index.codec.KNN920Codec.KNN920Codec
-org.opensearch.knn.index.codec.KNN940Codec.KNN940Codec
-org.opensearch.knn.index.codec.KNN950Codec.KNN950Codec
-org.opensearch.knn.index.codec.KNN990Codec.KNN990Codec
+org.opensearch.knn.index.codec.backward_codecs.KNN910Codec.KNN910Codec
+org.opensearch.knn.index.codec.backward_codecs.KNN920Codec.KNN920Codec
+org.opensearch.knn.index.codec.backward_codecs.KNN940Codec.KNN940Codec
+org.opensearch.knn.index.codec.backward_codecs.KNN950Codec.KNN950Codec
+org.opensearch.knn.index.codec.backward_codecs.KNN990Codec.KNN990Codec
 org.opensearch.knn.index.codec.backward_codecs.KNN9120Codec.KNN9120Codec
 org.opensearch.knn.index.codec.backward_codecs.KNN10010Codec.KNN10010Codec
 org.opensearch.knn.index.codec.backward_codecs.KNN1030Codec.KNN1030Codec
