+++ src/test/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -1,3 +1,3 @@
 org.opensearch.knn.index.codec.CustomCodec
 org.opensearch.knn.index.codec.CustomCodecNoStoredFields
-org.opensearch.knn.index.codec.KNN990Codec.UnitTestCodec
+org.opensearch.knn.index.codec.util.UnitTestCodec
